# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "KAFKA UI",
    "version": "0.1",
    "title": "Swagger of Kafka",
    "contact": {
      "email": "ferdi.tatlisu@gmail.com"
    }
  },
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/search": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Search any keyword in key or value",
        "description": "Search any keyword in key or value",
        "operationId": "search",
        "parameters": [
          {
            "name": "topicName",
            "in": "query",
            "description": "topic name",
            "required": true,
            "type": "string",
            "default": "product.liwe.heimdall-service.listing.0"
          },
          {
            "name": "value",
            "in": "query",
            "description": "search value",
            "required": true,
            "type": "string",
            "default": "4090d6f6f9b6663b6e4ef5d2996ae286"
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "startDate as ms",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": "4090d6f6f9b6663b6e4ef5d2996ae286"
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "endDate as ms",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": "4090d6f6f9b6663b6e4ef5d2996ae286"
          },
          {
            "name": "valueType",
            "in": "query",
            "description": "search in key(1), value(2) or header(4). Sum values of multiple selection",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": "4090d6f6f9b6663b6e4ef5d2996ae286"
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SearchResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "kafka"
        ],
        "summary": "Delete search data",
        "description": "Delete search data",
        "operationId": "delete-search",
        "parameters": [
          {
            "name": "topicName",
            "in": "query",
            "description": "topic name",
            "required": true,
            "type": "string",
            "default": "product.liwe.heimdall-service.listing.0"
          },
          {
            "name": "value",
            "in": "query",
            "description": "search value",
            "required": true,
            "type": "string",
            "default": "4090d6f6f9b6663b6e4ef5d2996ae286"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/topics/{topic}/groupIds/{groupId}/lag": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Get laf of topic with groupId",
        "description": "Get laf of topic with groupId",
        "operationId": "get-topic-lag",
        "parameters": [
          {
            "name": "topic",
            "in": "path",
            "description": "Topic Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "groupId",
            "in": "path",
            "description": "GroupId",
            "required": true,
            "type": "string"
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/topics": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Get all topics",
        "description": "Get all topics",
        "operationId": "topics",
        "parameters": [
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "kafka"
        ],
        "summary": "Create a new topic",
        "description": "Create a new topic",
        "operationId": "post-topics",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create topic",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTopicRequest"
            }
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "kafka"
        ],
        "summary": "Remove topic",
        "description": "Remove topic",
        "operationId": "delete-topic",
        "parameters": [
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          },
          {
            "name": "topic",
            "in": "query",
            "description": "topic",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/consumers": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Get all consumer groups",
        "description": "Get all consumer groups",
        "operationId": "consumers",
        "parameters": [
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "kafka"
        ],
        "summary": "Remove consumer group",
        "description": "Remove consumer group",
        "operationId": "delete-consumer-group",
        "parameters": [
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          },
          {
            "name": "id",
            "in": "query",
            "description": "group id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/consumer-group-by-topic": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Get all consumer groups via topic",
        "description": "Get all consumer groups via topic",
        "operationId": "consumer-group-by-topic",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "description": "topic name",
            "required": true,
            "type": "string",
            "default": "product.liwe.heimdall-service.listing.0"
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/consumer-group-by-group-id": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Get all consumer groups via group id",
        "description": "Get all consumer groups via group id",
        "operationId": "consumer-group-by-group-id",
        "parameters": [
          {
            "name": "group_id",
            "in": "query",
            "description": "group id",
            "required": true,
            "type": "string",
            "default": "product.liwe.heimdall-service.ruleCreatedEvent.high"
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/get-top-messages": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Get messages of topic",
        "description": "Get messages of topic",
        "operationId": "get-top-messages",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "description": "topic",
            "required": true,
            "type": "string",
            "default": "product.liwe.heimdall-service.listing.0"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 10
          },
          {
            "name": "partition",
            "in": "query",
            "description": "partition",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/get-message": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Get one message of topic",
        "description": "Get one message of topic",
        "operationId": "get-message",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "description": "topic",
            "required": true,
            "type": "string",
            "default": "product.liwe.heimdall-service.listing.0"
          },
          {
            "name": "partition",
            "in": "query",
            "description": "partition",
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "offset",
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/get-topic-info": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Summarize of topic",
        "description": "Summarize of topic",
        "operationId": "get-topic-info",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "description": "topic",
            "required": true,
            "type": "string",
            "default": "product.liwe.heimdall-service.listing.0"
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/get-consumer-info": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Summarize of consumer group",
        "description": "Summarize of consumer group",
        "operationId": "get-consumer-info",
        "parameters": [
          {
            "name": "group_id",
            "in": "query",
            "description": "group_id",
            "required": true,
            "type": "string",
            "default": "product.liwe.heimdall-service.ruleCreatedEvent.high"
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/get-topic-configuration": {
      "get": {
        "tags": [
          "kafka"
        ],
        "summary": "Get all configuration of topic",
        "description": "Get all configuration of topic",
        "operationId": "get-topic-configuration",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "description": "topic",
            "required": true,
            "type": "string",
            "default": "product.liwe.heimdall-service.listing.0"
          },
          {
            "name": "kafka-id",
            "in": "header",
            "description": "kafka-id",
            "required": true,
            "type": "integer",
            "format": "int64",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
  "definitions": {
    "SearchResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "messages": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
...
    "CreateTopicRequest": {
      "type": "object",
      "required": [
        "topic",
        "partitionCount",
        "retentionMs"
      ],
      "properties": {
        "topic": {
          "type": "string"
        },
        "partitionCount": {
          "type": "integer",
          "format": "int64"
        },
        "retentionMs": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
  },
...
}